# backend/openapi.json
{
  "components": {
    "schemas": {
      "ConversationDetailOut": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "last_message_at": {
            "format": "date-time",
            "title": "Last Message At",
            "type": "string"
          },
          "messages": {
            "items": {
              "$ref": "#/components/schemas/MessageOut"
            },
            "title": "Messages",
            "type": "array"
          },
          "model": {
            "title": "Model",
            "type": "string"
          },
          "project_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Project Id"
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          }
        },
        "required": [
          "id",
          "title",
          "model",
          "project_id",
          "created_at",
          "last_message_at",
          "messages"
        ],
        "title": "ConversationDetailOut",
        "type": "object"
      },
      "ConversationIn": {
        "properties": {
          "model": {
            "default": "talvrin-go",
            "title": "Model",
            "type": "string"
          },
          "project_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Project Id"
          },
          "title": {
            "anyOf": [
              {
                "maxLength": 400,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          }
        },
        "title": "ConversationIn",
        "type": "object"
      },
      "ConversationOut": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "last_message_at": {
            "format": "date-time",
            "title": "Last Message At",
            "type": "string"
          },
          "model": {
            "title": "Model",
            "type": "string"
          },
          "project_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Project Id"
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          }
        },
        "required": [
          "id",
          "title",
          "model",
          "project_id",
          "created_at",
          "last_message_at"
        ],
        "title": "ConversationOut",
        "type": "object"
      },
      "ConversationPatch": {
        "properties": {
          "project_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Project Id"
          },
          "title": {
            "anyOf": [
              {
                "maxLength": 400,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          }
        },
        "title": "ConversationPatch",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "MessageIn": {
        "description": "A message from the person typing.\n\nThere is deliberately no `role` field. A client that could post an\nASSISTANT message could fabricate something Talvrin never said \u2014 and then\nshare the link, where it would read as the platform's own output. For a\nproduct whose whole claim is sourced research and no recommendations, that\nis the forgery worth designing out rather than policing later. Assistant\nand system turns are written by the server, through the service layer.",
        "properties": {
          "content": {
            "maxLength": 100000,
            "minLength": 1,
            "title": "Content",
            "type": "string"
          }
        },
        "required": [
          "content"
        ],
        "title": "MessageIn",
        "type": "object"
      },
      "MessageOut": {
        "properties": {
          "content": {
            "title": "Content",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "role": {
            "title": "Role",
            "type": "string"
          },
          "seq": {
            "title": "Seq",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "seq",
          "role",
          "content",
          "created_at"
        ],
        "title": "MessageOut",
        "type": "object"
      },
...
      "ProjectIn": {
        "properties": {
          "name": {
            "maxLength": 400,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "name"
        ],
        "title": "ProjectIn",
        "type": "object"
      },
...
      "ShareIn": {
        "properties": {
          "expires_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expires At"
          }
        },
        "title": "ShareIn",
        "type": "object"
      },
      "ShareOut": {
        "properties": {
          "expires_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expires At"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "shared_up_to_seq": {
            "title": "Shared Up To Seq",
            "type": "integer"
          },
          "token": {
            "title": "Token",
            "type": "string"
          }
        },
        "required": [
          "id",
          "token",
          "shared_up_to_seq",
          "expires_at"
        ],
        "title": "ShareOut",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Talvrin API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/chats": {
      "get": {
        "operationId": "list_chats_api_v1_chats_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ConversationOut"
                  },
                  "title": "Response List Chats Api V1 Chats Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List Chats",
        "tags": [
          "research"
        ]
      },
      "post": {
        "operationId": "create_chat_api_v1_chats_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConversationIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConversationOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Chat",
        "tags": [
          "research"
        ]
      }
    },
    "/api/v1/chats/{chat_id}": {
      "delete": {
        "operationId": "delete_chat_api_v1_chats__chat_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "chat_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Chat Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Chat",
        "tags": [
          "research"
        ]
      },
      "get": {
        "operationId": "get_chat_api_v1_chats__chat_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "chat_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Chat Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConversationDetailOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Chat",
        "tags": [
          "research"
        ]
      },
      "patch": {
        "operationId": "patch_chat_api_v1_chats__chat_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "chat_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Chat Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConversationPatch"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Patch Chat",
        "tags": [
          "research"
        ]
      }
    },
    "/api/v1/chats/{chat_id}/messages": {
      "post": {
        "operationId": "append_message_api_v1_chats__chat_id__messages_post",
        "parameters": [
          {
            "in": "path",
            "name": "chat_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Chat Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MessageIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Append Message",
        "tags": [
          "research"
        ]
      }
    },
    "/api/v1/chats/{chat_id}/share": {
      "post": {
        "operationId": "create_share_api_v1_chats__chat_id__share_post",
        "parameters": [
          {
            "in": "path",
            "name": "chat_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Chat Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ShareIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ShareOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Share",
        "tags": [
          "research"
        ]
      }
    },
    "/api/v1/chats/{chat_id}/shares": {
      "get": {
        "operationId": "list_shares_api_v1_chats__chat_id__shares_get",
        "parameters": [
          {
            "in": "path",
            "name": "chat_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Chat Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ShareSummaryOut"
                  },
                  "title": "Response List Shares Api V1 Chats  Chat Id  Shares Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Shares",
        "tags": [
          "research"
        ]
      }
    },
...
    "/api/v1/projects": {
      "get": {
        "operationId": "list_projects_api_v1_projects_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ProjectOut"
                  },
                  "title": "Response List Projects Api V1 Projects Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List Projects",
        "tags": [
          "research"
        ]
      },
      "post": {
        "operationId": "create_project_api_v1_projects_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProjectIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Project",
        "tags": [
          "research"
        ]
      }
    },
    "/api/v1/projects/{project_id}": {
      "delete": {
        "operationId": "delete_project_api_v1_projects__project_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Project Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Project",
        "tags": [
          "research"
        ]
      },
      "patch": {
        "operationId": "rename_project_api_v1_projects__project_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Project Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProjectIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Rename Project",
        "tags": [
          "research"
        ]
      }
    },
    "/api/v1/shared/{token}": {
      "get": {
        "description": "Public. Deliberately takes no identity: the token is the whole\nauthorisation, and signing in must not change what a link shows.",
        "operationId": "read_shared_api_v1_shared__token__get",
        "parameters": [
          {
            "in": "path",
            "name": "token",
            "required": true,
            "schema": {
              "title": "Token",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SharedViewOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Read Shared",
        "tags": [
          "research"
        ]
      }
    },
    "/api/v1/shares/{share_id}": {
      "delete": {
        "operationId": "revoke_share_api_v1_shares__share_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "share_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Share Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Revoke Share",
        "tags": [
          "research"
        ]
      }
    },
...
  }
}